Added CPU and Fixed Bug While Playing
CPU:**********
The CPU was implimented under the Board class.
It has its own method called:

cpuMove(self)

The cpuMove will look for either a place to win if ther are 2 'O' in a row
or a place to block if there are two 'X' in a row. It utilizes the other method called
findBlock(self, line) which will return a number of the space to play in depending on the
line that is given to it.

BUG FIX**********
There was a bug in the program that while playing if the user entered in anything other than a number
it would exit the program. The bug was fixed with a simple if statement making sure the user entered in a
numerical value before trying to place a move on the gameboard.